+++ lib/papa/helpers.rb
@@ -1,0 +1,4 @@
+module Papa
+  class Helpers
+  end
+end