+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.0
 * Used open3 for easier output and exit status handling
 * Added more descriptive output messages
+* Prompt vi when no branches are specified in add
 
 ## 0.2.0
 * Added `integration [start]`